Set up a C development environment
This does a few things:

* conditionally loads the Ruby style settings defined by the core team
if there is a CRuby checkout in the place I normally keep
it (~/src/ruby)

* Ensures lsp-mode is installed and active for all C mode files.

* Makes sure yasnippet mode is activated

* Overrides the default keymap (super-l) to (Ctrl-c l) because I use
super as meta.

Note that this will start clangd automatically in the background. To
make this work it needs to know how to build your project. Generate a
compile_commands.json in your project dir using the tool bear.

I did:

  bear -- make -j13