Switch package managers, use Org mode for Init file
So there's a fair amount of stuff bundled into this commit, which is
kinda bad, but the main 2 things are:

- Switch from Emacs built-in package manager to straight.el - for all
of the reasons outlined here:
https://github.com/radian-software/straight.el#advantages-of-straightel
- use org-babel to generate my emacs init file (init.el) dynamically
from a template init.org.

This will allow me to properly document my emacs config in a literate
programming fashion, which will certainly aid my memory, and hopefully
make this config more useful to others in the process.

During my testing I ran into the issues described in this thread:
https://www.reddit.com/r/emacs/comments/qcj33a/problem_and_workaround_with_orgmode_function/

Namely, that the version mismatch between the system provided org-mode
and the requirements of some of the various org packages that
straight.el pulls in was causing function not found errors.

To mitigate this I've introduced an early-init file that bootstraps
straight.el, configures use-package to always use straight.el, install
org-mode and then process my init.org file.

Because the early-init file is executed before the GUI is running, an
initial install on a new machine is kind of awkward - the machine
appears to hang while org-mode is installed.