tangle README.org into init.el not README.el
Using org-bable-load-file here loads the generated file immediately. This is
mostly fine, except it's loaded in the contest of the early-init
environment. This means that relying on (window-system) isn't going to work
because it hasn't been defined yet.

And because org-bable-load-file doesn't tangle into init.el we have no actual
init script that will be loaded at a later point.

Instead we can just use org-babel-tangle-file in the early init to explicitly
generate an init.el from our README.org - then the generated init.el will be
loaded implicitly as part of the rest of Emacs startup. Which will result in all
the context we need being available.